split Executor impl into separate file from API
