+++ src/execution/Executor.ts
@@ -3,9 +3,7 @@ import { inspect } from '../jsutils/inspect.js';
 import { invariant } from '../jsutils/invariant.js';
 import { isAsyncIterable } from '../jsutils/isAsyncIterable.js';
 import { isIterableObject } from '../jsutils/isIterableObject.js';
-import { isObjectLike } from '../jsutils/isObjectLike.js';
 import { isPromise } from '../jsutils/isPromise.js';
-import type { Maybe } from '../jsutils/Maybe.js';
 import { memoize3 } from '../jsutils/memoize3.js';
 import type { ObjMap } from '../jsutils/ObjMap.js';
 import type { Path } from '../jsutils/Path.js';
@@ -18,13 +16,11 @@ import { GraphQLError } from '../error/GraphQLError.js';
 import { locatedError } from '../error/locatedError.js';
 
 import type {
-  DocumentNode,
   FieldNode,
   FragmentDefinitionNode,
   OperationDefinitionNode,
 } from '../language/ast.js';
 import { OperationTypeNode } from '../language/ast.js';
-import { Kind } from '../language/kinds.js';
 
 import type {
   GraphQLAbstractType,
@@ -46,7 +42,6 @@ import {
 } from '../type/definition.js';
 import { GraphQLStreamDirective } from '../type/directives.js';
 import type { GraphQLSchema } from '../type/schema.js';
-import { assertValidSchema } from '../type/validate.js';
 
 import {
   AbortSignalListener,
@@ -65,9 +60,7 @@ import {
   collectFields,
   collectSubfields as _collectSubfields,
 } from './collectFields.js';
-import { getVariableSignature } from './getVariableSignature.js';
 import { buildIncrementalResponse } from './IncrementalPublisher.js';
-import { mapAsyncIterable } from './mapAsyncIterable.js';
 import type {
   CancellableStreamRecord,
   CompletedExecutionGroup,
@@ -81,12 +74,7 @@ import type {
 } from './types.js';
 import { DeferredFragmentRecord } from './types.js';
 import type { VariableValues } from './values.js';
-import {
-  experimentalGetArgumentValues,
-  getArgumentValues,
-  getDirectiveValues,
-  getVariableValues,
-} from './values.js';
+import { experimentalGetArgumentValues, getDirectiveValues } from './values.js';
 
 /* eslint-disable @typescript-eslint/max-params */
 // This file contains a lot of such errors but we plan to refactor it anyway
@@ -178,27 +166,7 @@ interface IncrementalContext {
   deferUsageSet?: DeferUsageSet | undefined;
 }
 
-export interface ExecutionArgs {
-  schema: GraphQLSchema;
-  document: DocumentNode;
-  rootValue?: unknown;
-  contextValue?: unknown;
-  variableValues?: Maybe<{ readonly [variable: string]: unknown }>;
-  operationName?: Maybe<string>;
-  fieldResolver?: Maybe<GraphQLFieldResolver<any, any>>;
-  typeResolver?: Maybe<GraphQLTypeResolver<any, any>>;
-  subscribeFieldResolver?: Maybe<GraphQLFieldResolver<any, any>>;
-  perEventExecutor?: Maybe<
-    (
-      validatedExecutionArgs: ValidatedExecutionArgs,
-    ) => PromiseOrValue<ExecutionResult>
-  >;
-  enableEarlyExecution?: Maybe<boolean>;
-  hideSuggestions?: Maybe<boolean>;
-  abortSignal?: Maybe<AbortSignal>;
-}
-
-export interface StreamUsage {
+interface StreamUsage {
   label: string | undefined;
   initialCount: number;
   fieldDetailsList: FieldDetailsList;
@@ -209,112 +177,7 @@ interface GraphQLWrappedResult<T> {
   incrementalDataRecords: Array<IncrementalDataRecord> | undefined;
 }
 
-const UNEXPECTED_EXPERIMENTAL_DIRECTIVES =
-  'The provided schema unexpectedly contains experimental directives (@defer or @stream). These directives may only be utilized if experimental execution features are explicitly enabled.';
-
-const UNEXPECTED_MULTIPLE_PAYLOADS =
-  'Executing this GraphQL operation would unexpectedly produce multiple payloads (due to @defer or @stream directive)';
-
-/**
- * Implements the "Executing requests" section of the GraphQL specification.
- *
- * Returns either a synchronous ExecutionResult (if all encountered resolvers
- * are synchronous), or a Promise of an ExecutionResult that will eventually be
- * resolved and never rejected.
- *
- * If the arguments to this function do not result in a legal execution context,
- * a GraphQLError will be thrown immediately explaining the invalid input.
- *
- * This function does not support incremental delivery (`@defer` and `@stream`).
- * If an operation which would defer or stream data is executed with this
- * function, it will throw or return a rejected promise.
- * Use `experimentalExecuteIncrementally` if you want to support incremental
- * delivery.
- */
-export function execute(args: ExecutionArgs): PromiseOrValue<ExecutionResult> {
-  if (args.schema.getDirective('defer') || args.schema.getDirective('stream')) {
-    throw new Error(UNEXPECTED_EXPERIMENTAL_DIRECTIVES);
-  }
-
-  const result = experimentalExecuteIncrementally(args);
-  // Multiple payloads could be encountered if the operation contains @defer or
-  // @stream directives and is not validated prior to execution
-  return ensureSinglePayload(result);
-}
-
-function ensureSinglePayload(
-  result: PromiseOrValue<
-    ExecutionResult | ExperimentalIncrementalExecutionResults
-  >,
-): PromiseOrValue<ExecutionResult> {
-  if (isPromise(result)) {
-    return result.then((resolved) => {
-      if ('initialResult' in resolved) {
-        throw new Error(UNEXPECTED_MULTIPLE_PAYLOADS);
-      }
-      return resolved;
-    });
-  }
-  if ('initialResult' in result) {
-    throw new Error(UNEXPECTED_MULTIPLE_PAYLOADS);
-  }
-  return result;
-}
-
-/**
- * Implements the "Executing requests" section of the GraphQL specification,
- * including `@defer` and `@stream` as proposed in
- * https://github.com/graphql/graphql-spec/pull/742
- *
- * This function returns a Promise of an ExperimentalIncrementalExecutionResults
- * object. This object either consists of a single ExecutionResult, or an
- * object containing an `initialResult` and a stream of `subsequentResults`.
- *
- * If the arguments to this function do not result in a legal execution context,
- * a GraphQLError will be thrown immediately explaining the invalid input.
- */
-export function experimentalExecuteIncrementally(
-  args: ExecutionArgs,
-): PromiseOrValue<ExecutionResult | ExperimentalIncrementalExecutionResults> {
-  // If a valid execution context cannot be created due to incorrect arguments,
-  // a "Response" with only errors is returned.
-  const validatedExecutionArgs = validateExecutionArgs(args);
-
-  // Return early errors if execution context failed.
-  if (!('schema' in validatedExecutionArgs)) {
-    return { errors: validatedExecutionArgs };
-  }
-
-  return experimentalExecuteQueryOrMutationOrSubscriptionEvent(
-    validatedExecutionArgs,
-  );
-}
-
-/**
- * Implements the "Executing operations" section of the spec.
- *
- * Returns a Promise that will eventually resolve to the data described by
- * The "Response" section of the GraphQL specification.
- *
- * If errors are encountered while executing a GraphQL field, only that
- * field and its descendants will be omitted, and sibling fields will still
- * be executed. An execution which encounters errors will still result in a
- * resolved Promise.
- *
- * Errors from sub-fields of a NonNull type may propagate to the top level,
- * at which point we still log the error and null the parent field, which
- * in this case is the entire response.
- */
-export function executeQueryOrMutationOrSubscriptionEvent(
-  validatedExecutionArgs: ValidatedExecutionArgs,
-): PromiseOrValue<ExecutionResult> {
-  const result = experimentalExecuteQueryOrMutationOrSubscriptionEvent(
-    validatedExecutionArgs,
-  );
-  return ensureSinglePayload(result);
-}
-
-export function experimentalExecuteQueryOrMutationOrSubscriptionEvent(
+export function executeQueryOrMutationOrSubscriptionEventImpl(
   validatedExecutionArgs: ValidatedExecutionArgs,
 ): PromiseOrValue<ExecutionResult | ExperimentalIncrementalExecutionResults> {
   const abortSignal = validatedExecutionArgs.abortSignal;
@@ -416,136 +279,6 @@ function buildDataResponse(
     errors,
     incrementalDataRecords,
   );
-}
-
-/**
- * Also implements the "Executing requests" section of the GraphQL specification.
- * However, it guarantees to complete synchronously (or throw an error) assuming
- * that all field resolvers are also synchronous.
- */
-export function executeSync(args: ExecutionArgs): ExecutionResult {
-  const result = experimentalExecuteIncrementally(args);
-
-  // Assert that the execution was synchronous.
-  if (isPromise(result) || 'initialResult' in result) {
-    throw new Error('GraphQL execution failed to complete synchronously.');
-  }
-
-  return result;
-}
-
-/**
- * Constructs a ExecutionContext object from the arguments passed to
- * execute, which we will pass throughout the other execution methods.
- *
- * Throws a GraphQLError if a valid execution context cannot be created.
- *
- * TODO: consider no longer exporting this function
- * @internal
- */
-export function validateExecutionArgs(
-  args: ExecutionArgs,
-): ReadonlyArray<GraphQLError> | ValidatedExecutionArgs {
-  const {
-    schema,
-    document,
-    rootValue,
-    contextValue,
-    variableValues: rawVariableValues,
-    operationName,
-    fieldResolver,
-    typeResolver,
-    subscribeFieldResolver,
-    perEventExecutor,
-    enableEarlyExecution,
-    abortSignal,
-  } = args;
-
-  if (abortSignal?.aborted) {
-    return [locatedError(abortSignal.reason, undefined)];
-  }
-
-  // If the schema used for execution is invalid, throw an error.
-  assertValidSchema(schema);
-
-  let operation: OperationDefinitionNode | undefined;
-  const fragmentDefinitions: ObjMap<FragmentDefinitionNode> =
-    Object.create(null);
-  const fragments: ObjMap<FragmentDetails> = Object.create(null);
-  for (const definition of document.definitions) {
-    switch (definition.kind) {
-      case Kind.OPERATION_DEFINITION:
-        if (operationName == null) {
-          if (operation !== undefined) {
-            return [
-              new GraphQLError(
-                'Must provide operation name if query contains multiple operations.',
-              ),
-            ];
-          }
-          operation = definition;
-        } else if (definition.name?.value === operationName) {
-          operation = definition;
-        }
-        break;
-      case Kind.FRAGMENT_DEFINITION: {
-        fragmentDefinitions[definition.name.value] = definition;
-        let variableSignatures;
-        if (definition.variableDefinitions) {
-          variableSignatures = Object.create(null);
-          for (const varDef of definition.variableDefinitions) {
-            const signature = getVariableSignature(schema, varDef);
-            variableSignatures[signature.name] = signature;
-          }
-        }
-        fragments[definition.name.value] = { definition, variableSignatures };
-        break;
-      }
-      default:
-      // ignore non-executable definitions
-    }
-  }
-
-  if (!operation) {
-    if (operationName != null) {
-      return [new GraphQLError(`Unknown operation named "${operationName}".`)];
-    }
-    return [new GraphQLError('Must provide an operation.')];
-  }
-
-  const variableDefinitions = operation.variableDefinitions ?? [];
-  const hideSuggestions = args.hideSuggestions ?? false;
-
-  const variableValuesOrErrors = getVariableValues(
-    schema,
-    variableDefinitions,
-    rawVariableValues ?? {},
-    {
-      maxErrors: 50,
-      hideSuggestions,
-    },
-  );
-
-  if (variableValuesOrErrors.errors) {
-    return variableValuesOrErrors.errors;
-  }
-
-  return {
-    schema,
-    fragmentDefinitions,
-    fragments,
-    rootValue,
-    contextValue,
-    operation,
-    variableValues: variableValuesOrErrors.variableValues,
-    fieldResolver: fieldResolver ?? defaultFieldResolver,
-    typeResolver: typeResolver ?? defaultTypeResolver,
-    subscribeFieldResolver: subscribeFieldResolver ?? defaultFieldResolver,
-    perEventExecutor: perEventExecutor ?? executeSubscriptionEvent,
-    enableEarlyExecution: enableEarlyExecution === true,
-    hideSuggestions,
-    abortSignal: args.abortSignal ?? undefined,
-  };
 }
 
 function executeRootExecutionPlan(
@@ -1998,346 +1731,6 @@ function buildSubExecutionPlan(
     originalGroupedFieldSet as unknown as { _executionPlan: ExecutionPlan }
   )._executionPlan = executionPlan;
   return executionPlan;
-}
-
-/**
- * If a resolveType function is not given, then a default resolve behavior is
- * used which attempts two strategies:
- *
- * First, See if the provided value has a `__typename` field defined, if so, use
- * that value as name of the resolved type.
- *
- * Otherwise, test each possible type for the abstract type by calling
- * isTypeOf for the object being coerced, returning the first type that matches.
- */
-export const defaultTypeResolver: GraphQLTypeResolver<unknown, unknown> =
-  function (value, contextValue, info, abstractType) {
-    // First, look for `__typename`.
-    if (isObjectLike(value) && typeof value.__typename === 'string') {
-      return value.__typename;
-    }
-
-    // Otherwise, test each possible type.
-    const possibleTypes = info.schema.getPossibleTypes(abstractType);
-    const promisedIsTypeOfResults = [];
-
-    for (let i = 0; i < possibleTypes.length; i++) {
-      const type = possibleTypes[i];
-
-      if (type.isTypeOf) {
-        const isTypeOfResult = type.isTypeOf(value, contextValue, info);
-
-        if (isPromise(isTypeOfResult)) {
-          promisedIsTypeOfResults[i] = isTypeOfResult;
-        } else if (isTypeOfResult) {
-          if (promisedIsTypeOfResults.length > 0) {
-            Promise.all(promisedIsTypeOfResults).then(undefined, () => {
-              /* ignore errors */
-            });
-          }
-
-          return type.name;
-        }
-      }
-    }
-
-    if (promisedIsTypeOfResults.length) {
-      return Promise.all(promisedIsTypeOfResults).then((isTypeOfResults) => {
-        for (let i = 0; i < isTypeOfResults.length; i++) {
-          if (isTypeOfResults[i]) {
-            return possibleTypes[i].name;
-          }
-        }
-      });
-    }
-  };
-
-/**
- * If a resolve function is not given, then a default resolve behavior is used
- * which takes the property of the source object of the same name as the field
- * and returns it as the result, or if it's a function, returns the result
- * of calling that function while passing along args and context value.
- */
-export const defaultFieldResolver: GraphQLFieldResolver<unknown, unknown> =
-  function (source: any, args, contextValue, info, abortSignal) {
-    // ensure source is a value for which property access is acceptable.
-    if (isObjectLike(source) || typeof source === 'function') {
-      const property = source[info.fieldName];
-      if (typeof property === 'function') {
-        return source[info.fieldName](args, contextValue, info, abortSignal);
-      }
-      return property;
-    }
-  };
-
-/**
- * Implements the "Subscribe" algorithm described in the GraphQL specification.
- *
- * Returns a Promise which resolves to either an AsyncIterator (if successful)
- * or an ExecutionResult (error). The promise will be rejected if the schema or
- * other arguments to this function are invalid, or if the resolved event stream
- * is not an async iterable.
- *
- * If the client-provided arguments to this function do not result in a
- * compliant subscription, a GraphQL Response (ExecutionResult) with descriptive
- * errors and no data will be returned.
- *
- * If the source stream could not be created due to faulty subscription resolver
- * logic or underlying systems, the promise will resolve to a single
- * ExecutionResult containing `errors` and no `data`.
- *
- * If the operation succeeded, the promise resolves to an AsyncIterator, which
- * yields a stream of ExecutionResults representing the response stream.
- *
- * This function does not support incremental delivery (`@defer` and `@stream`).
- * If an operation which would defer or stream data is executed with this
- * function, a field error will be raised at the location of the `@defer` or
- * `@stream` directive.
- *
- * Accepts an object with named arguments.
- */
-export function subscribe(
-  args: ExecutionArgs,
-): PromiseOrValue<
-  AsyncGenerator<ExecutionResult, void, void> | ExecutionResult
-> {
-  // If a valid execution context cannot be created due to incorrect arguments,
-  // a "Response" with only errors is returned.
-  const validatedExecutionArgs = validateExecutionArgs(args);
-
-  // Return early errors if execution context failed.
-  if (!('schema' in validatedExecutionArgs)) {
-    return { errors: validatedExecutionArgs };
-  }
-
-  const resultOrStream = createSourceEventStreamImpl(validatedExecutionArgs);
-
-  if (isPromise(resultOrStream)) {
-    return resultOrStream.then((resolvedResultOrStream) =>
-      mapSourceToResponse(validatedExecutionArgs, resolvedResultOrStream),
-    );
-  }
-
-  return mapSourceToResponse(validatedExecutionArgs, resultOrStream);
-}
-
-function mapSourceToResponse(
-  validatedExecutionArgs: ValidatedExecutionArgs,
-  resultOrStream: ExecutionResult | AsyncIterable<unknown>,
-): AsyncGenerator<ExecutionResult, void, void> | ExecutionResult {
-  if (!isAsyncIterable(resultOrStream)) {
-    return resultOrStream;
-  }
-
-  const abortSignal = validatedExecutionArgs.abortSignal;
-  const abortSignalListener = abortSignal
-    ? new AbortSignalListener(abortSignal)
-    : undefined;
-
-  // For each payload yielded from a subscription, map it over the normal
-  // GraphQL `execute` function, with `payload` as the rootValue.
-  // This implements the "MapSourceToResponseEvent" algorithm described in
-  // the GraphQL specification..
-  return mapAsyncIterable(
-    abortSignalListener
-      ? cancellableIterable(resultOrStream, abortSignalListener)
-      : resultOrStream,
-    (payload: unknown) => {
-      const perEventExecutionArgs: ValidatedExecutionArgs = {
-        ...validatedExecutionArgs,
-        rootValue: payload,
-      };
-      return validatedExecutionArgs.perEventExecutor(perEventExecutionArgs);
-    },
-    () => abortSignalListener?.disconnect(),
-  );
-}
-
-export function executeSubscriptionEvent(
-  validatedExecutionArgs: ValidatedExecutionArgs,
-): PromiseOrValue<ExecutionResult> {
-  return executeQueryOrMutationOrSubscriptionEvent(validatedExecutionArgs);
-}
-
-/**
- * Implements the "CreateSourceEventStream" algorithm described in the
- * GraphQL specification, resolving the subscription source event stream.
- *
- * Returns a Promise which resolves to either an AsyncIterable (if successful)
- * or an ExecutionResult (error). The promise will be rejected if the schema or
- * other arguments to this function are invalid, or if the resolved event stream
- * is not an async iterable.
- *
- * If the client-provided arguments to this function do not result in a
- * compliant subscription, a GraphQL Response (ExecutionResult) with
- * descriptive errors and no data will be returned.
- *
- * If the the source stream could not be created due to faulty subscription
- * resolver logic or underlying systems, the promise will resolve to a single
- * ExecutionResult containing `errors` and no `data`.
- *
- * If the operation succeeded, the promise resolves to the AsyncIterable for the
- * event stream returned by the resolver.
- *
- * A Source Event Stream represents a sequence of events, each of which triggers
- * a GraphQL execution for that event.
- *
- * This may be useful when hosting the stateful subscription service in a
- * different process or machine than the stateless GraphQL execution engine,
- * or otherwise separating these two steps. For more on this, see the
- * "Supporting Subscriptions at Scale" information in the GraphQL specification.
- */
-export function createSourceEventStream(
-  args: ExecutionArgs,
-): PromiseOrValue<AsyncIterable<unknown> | ExecutionResult> {
-  // If a valid execution context cannot be created due to incorrect arguments,
-  // a "Response" with only errors is returned.
-  const validatedExecutionArgs = validateExecutionArgs(args);
-
-  // Return early errors if execution context failed.
-  if (!('schema' in validatedExecutionArgs)) {
-    return { errors: validatedExecutionArgs };
-  }
-
-  return createSourceEventStreamImpl(validatedExecutionArgs);
-}
-
-function createSourceEventStreamImpl(
-  validatedExecutionArgs: ValidatedExecutionArgs,
-): PromiseOrValue<AsyncIterable<unknown> | ExecutionResult> {
-  try {
-    const eventStream = executeSubscription(validatedExecutionArgs);
-    if (isPromise(eventStream)) {
-      return eventStream.then(undefined, (error: unknown) => ({
-        errors: [error as GraphQLError],
-      }));
-    }
-
-    return eventStream;
-  } catch (error) {
-    return { errors: [error] };
-  }
-}
-
-function executeSubscription(
-  validatedExecutionArgs: ValidatedExecutionArgs,
-): PromiseOrValue<AsyncIterable<unknown>> {
-  const {
-    schema,
-    fragments,
-    rootValue,
-    contextValue,
-    operation,
-    variableValues,
-    hideSuggestions,
-    abortSignal,
-  } = validatedExecutionArgs;
-
-  const rootType = schema.getSubscriptionType();
-  if (rootType == null) {
-    throw new GraphQLError(
-      'Schema is not configured to execute subscription operation.',
-      { nodes: operation },
-    );
-  }
-
-  const { groupedFieldSet } = collectFields(
-    schema,
-    fragments,
-    variableValues,
-    rootType,
-    operation.selectionSet,
-    hideSuggestions,
-  );
-
-  const firstRootField = groupedFieldSet.entries().next().value as [
-    string,
-    FieldDetailsList,
-  ];
-  const [responseName, fieldDetailsList] = firstRootField;
-  const fieldName = fieldDetailsList[0].node.name.value;
-  const fieldDef = schema.getField(rootType, fieldName);
-
-  const fieldNodes = fieldDetailsList.map((fieldDetails) => fieldDetails.node);
-  if (!fieldDef) {
-    throw new GraphQLError(
-      `The subscription field "${fieldName}" is not defined.`,
-      { nodes: fieldNodes },
-    );
-  }
-
-  const path = addPath(undefined, responseName, rootType.name);
-  const info = buildResolveInfo(
-    validatedExecutionArgs,
-    fieldDef,
-    fieldNodes,
-    rootType,
-    path,
-  );
-
-  try {
-    // Implements the "ResolveFieldEventStream" algorithm from GraphQL specification.
-    // It differs from "ResolveFieldValue" due to providing a different `resolveFn`.
-
-    // Build a JS object of arguments from the field.arguments AST, using the
-    // variables scope to fulfill any variable references.
-    const args = getArgumentValues(
-      fieldDef,
-      fieldNodes[0],
-      variableValues,
-      hideSuggestions,
-    );
-
-    // Call the `subscribe()` resolver or the default resolver to produce an
-    // AsyncIterable yielding raw payloads.
-    const resolveFn =
-      fieldDef.subscribe ?? validatedExecutionArgs.subscribeFieldResolver;
-
-    // The resolve function's optional third argument is a context value that
-    // is provided to every resolve function within an execution. It is commonly
-    // used to represent an authenticated user, or request-specific caches.
-    const result = resolveFn(rootValue, args, contextValue, info, abortSignal);
-
-    if (isPromise(result)) {
-      const abortSignalListener = abortSignal
-        ? new AbortSignalListener(abortSignal)
-        : undefined;
-
-      const promise = abortSignalListener
-        ? cancellablePromise(result, abortSignalListener)
-        : result;
-      return promise.then(assertEventStream).then(
-        (resolved) => {
-          abortSignalListener?.disconnect();
-          return resolved;
-        },
-        (error: unknown) => {
-          abortSignalListener?.disconnect();
-          throw locatedError(error, fieldNodes, pathToArray(path));
-        },
-      );
-    }
-
-    return assertEventStream(result);
-  } catch (error) {
-    throw locatedError(error, fieldNodes, pathToArray(path));
-  }
-}
-
-function assertEventStream(result: unknown): AsyncIterable<unknown> {
-  if (result instanceof Error) {
-    throw result;
-  }
-
-  // Assert field returned an event stream, otherwise yield an error.
-  if (!isAsyncIterable(result)) {
-    throw new GraphQLError(
-      'Subscription field must return Async Iterable. ' +
-        `Received: ${inspect(result)}.`,
-    );
-  }
-
-  return result;
 }
 
 function collectExecutionGroups(
