+++ src/execution/__tests__/stream-test.ts
@@ -19,7 +19,7 @@ import {
 import { GraphQLID, GraphQLString } from '../../type/scalars.js';
 import { GraphQLSchema } from '../../type/schema.js';
 
-import { experimentalExecuteIncrementally } from '../Executor.js';
+import { experimentalExecuteIncrementally } from '../execute.js';
 import type {
   InitialIncrementalExecutionResult,
   SubsequentIncrementalExecutionResult,
